requirements.txt
asgiref==3.11.1
Django==5.2.11
gunicorn==25.1.0
packaging==26.0
psycopg2-binary==2.9.11
python-decouple==3.8
python-dotenv==1.2.1
sqlparse==0.5.5
